New files added in order to move to MVVM archtecture: HomeViewModel, RetrofitInstance. getCryptoPrice has been passed as parameter to HomeStartScreen, which needs that fun to obtain BTC value. getRetrofit has been moved.

+++ app/src/main/java/com/example/financesmanagementapp/model/database/BinanceApiBTCPriceResponse.kt
@@ -5,6 +5,6 @@ import com.google.gson.annotations.SerializedName
 data class BinanceApiBTCPriceResponse(
     @SerializedName("mins")var minutes:Int,
     @SerializedName("price")var btcCurrentPrice:String,
-    @SerializedName("closeTime")var closeTime:Int
+    @SerializedName("closeTime")var closeTime:Long
 )
